+++ src/OpenTelemetry.AutoInstrumentation.Native/assembly_redirection_net.h
@@ -32,16 +32,16 @@
         { _W("OpenTelemetry.Exporter.Prometheus.HttpListener"), {1, 0, 0, 0} }, \
         { _W("OpenTelemetry.Exporter.Zipkin"), {1, 0, 0, 0} }, \
         { _W("OpenTelemetry.Extensions.Propagators"), {1, 0, 0, 0} }, \
-        { _W("OpenTelemetry.Instrumentation.AspNetCore"), {1, 15, 2, 1013} }, \
-        { _W("OpenTelemetry.Instrumentation.EntityFrameworkCore"), {1, 15, 1, 1021} }, \
-        { _W("OpenTelemetry.Instrumentation.GrpcNetClient"), {1, 15, 1, 1023} }, \
-        { _W("OpenTelemetry.Instrumentation.Http"), {1, 15, 1, 1025} }, \
-        { _W("OpenTelemetry.Instrumentation.Process"), {1, 15, 1, 1028} }, \
+        { _W("OpenTelemetry.Instrumentation.AspNetCore"), {1, 16, 0, 1140} }, \
+        { _W("OpenTelemetry.Instrumentation.EntityFrameworkCore"), {1, 16, 0, 1129} }, \
+        { _W("OpenTelemetry.Instrumentation.GrpcNetClient"), {1, 16, 0, 1132} }, \
+        { _W("OpenTelemetry.Instrumentation.Http"), {1, 16, 0, 1135} }, \
+        { _W("OpenTelemetry.Instrumentation.Process"), {1, 16, 0, 1136} }, \
         { _W("OpenTelemetry.Instrumentation.Quartz"), {1, 15, 1, 1027} }, \
         { _W("OpenTelemetry.Instrumentation.Runtime"), {1, 15, 1, 1029} }, \
-        { _W("OpenTelemetry.Instrumentation.SqlClient"), {1, 15, 2, 1033} }, \
-        { _W("OpenTelemetry.Instrumentation.StackExchangeRedis"), {1, 15, 1, 1093} }, \
-        { _W("OpenTelemetry.Instrumentation.Wcf"), {1, 15, 1, 1049} }, \
+        { _W("OpenTelemetry.Instrumentation.SqlClient"), {1, 16, 0, 1138} }, \
+        { _W("OpenTelemetry.Instrumentation.StackExchangeRedis"), {1, 16, 0, 1137} }, \
+        { _W("OpenTelemetry.Instrumentation.Wcf"), {1, 16, 0, 1139} }, \
         { _W("OpenTelemetry.OpAmp.Client"), {0, 4, 0, 1087} }, \
         { _W("OpenTelemetry.Resources.Azure"), {1, 15, 1, 1031} }, \
         { _W("OpenTelemetry.Resources.Container"), {1, 15, 1, 1038} }, \
@@ -83,16 +83,16 @@
         { _W("OpenTelemetry.Exporter.Prometheus.HttpListener"), {1, 0, 0, 0} }, \
         { _W("OpenTelemetry.Exporter.Zipkin"), {1, 0, 0, 0} }, \
         { _W("OpenTelemetry.Extensions.Propagators"), {1, 0, 0, 0} }, \
-        { _W("OpenTelemetry.Instrumentation.AspNetCore"), {1, 15, 2, 1013} }, \
-        { _W("OpenTelemetry.Instrumentation.EntityFrameworkCore"), {1, 15, 1, 1021} }, \
-        { _W("OpenTelemetry.Instrumentation.GrpcNetClient"), {1, 15, 1, 1023} }, \
-        { _W("OpenTelemetry.Instrumentation.Http"), {1, 15, 1, 1025} }, \
-        { _W("OpenTelemetry.Instrumentation.Process"), {1, 15, 1, 1028} }, \
+        { _W("OpenTelemetry.Instrumentation.AspNetCore"), {1, 16, 0, 1140} }, \
+        { _W("OpenTelemetry.Instrumentation.EntityFrameworkCore"), {1, 16, 0, 1129} }, \
+        { _W("OpenTelemetry.Instrumentation.GrpcNetClient"), {1, 16, 0, 1132} }, \
+        { _W("OpenTelemetry.Instrumentation.Http"), {1, 16, 0, 1135} }, \
+        { _W("OpenTelemetry.Instrumentation.Process"), {1, 16, 0, 1136} }, \
         { _W("OpenTelemetry.Instrumentation.Quartz"), {1, 15, 1, 1027} }, \
         { _W("OpenTelemetry.Instrumentation.Runtime"), {1, 15, 1, 1029} }, \
-        { _W("OpenTelemetry.Instrumentation.SqlClient"), {1, 15, 2, 1033} }, \
-        { _W("OpenTelemetry.Instrumentation.StackExchangeRedis"), {1, 15, 1, 1093} }, \
-        { _W("OpenTelemetry.Instrumentation.Wcf"), {1, 15, 1, 1049} }, \
+        { _W("OpenTelemetry.Instrumentation.SqlClient"), {1, 16, 0, 1138} }, \
+        { _W("OpenTelemetry.Instrumentation.StackExchangeRedis"), {1, 16, 0, 1137} }, \
+        { _W("OpenTelemetry.Instrumentation.Wcf"), {1, 16, 0, 1139} }, \
         { _W("OpenTelemetry.OpAmp.Client"), {0, 4, 0, 1087} }, \
         { _W("OpenTelemetry.Resources.Azure"), {1, 15, 1, 1031} }, \
         { _W("OpenTelemetry.Resources.Container"), {1, 15, 1, 1038} }, \
@@ -134,16 +134,16 @@
         { _W("OpenTelemetry.Exporter.Prometheus.HttpListener"), {1, 0, 0, 0} }, \
         { _W("OpenTelemetry.Exporter.Zipkin"), {1, 0, 0, 0} }, \
         { _W("OpenTelemetry.Extensions.Propagators"), {1, 0, 0, 0} }, \
-        { _W("OpenTelemetry.Instrumentation.AspNetCore"), {1, 15, 2, 1013} }, \
-        { _W("OpenTelemetry.Instrumentation.EntityFrameworkCore"), {1, 15, 1, 1021} }, \
-        { _W("OpenTelemetry.Instrumentation.GrpcNetClient"), {1, 15, 1, 1023} }, \
-        { _W("OpenTelemetry.Instrumentation.Http"), {1, 15, 1, 1025} }, \
-        { _W("OpenTelemetry.Instrumentation.Process"), {1, 15, 1, 1028} }, \
+        { _W("OpenTelemetry.Instrumentation.AspNetCore"), {1, 16, 0, 1140} }, \
+        { _W("OpenTelemetry.Instrumentation.EntityFrameworkCore"), {1, 16, 0, 1129} }, \
+        { _W("OpenTelemetry.Instrumentation.GrpcNetClient"), {1, 16, 0, 1132} }, \
+        { _W("OpenTelemetry.Instrumentation.Http"), {1, 16, 0, 1135} }, \
+        { _W("OpenTelemetry.Instrumentation.Process"), {1, 16, 0, 1136} }, \
         { _W("OpenTelemetry.Instrumentation.Quartz"), {1, 15, 1, 1027} }, \
         { _W("OpenTelemetry.Instrumentation.Runtime"), {1, 15, 1, 1029} }, \
-        { _W("OpenTelemetry.Instrumentation.SqlClient"), {1, 15, 2, 1033} }, \
-        { _W("OpenTelemetry.Instrumentation.StackExchangeRedis"), {1, 15, 1, 1093} }, \
-        { _W("OpenTelemetry.Instrumentation.Wcf"), {1, 15, 1, 1049} }, \
+        { _W("OpenTelemetry.Instrumentation.SqlClient"), {1, 16, 0, 1138} }, \
+        { _W("OpenTelemetry.Instrumentation.StackExchangeRedis"), {1, 16, 0, 1137} }, \
+        { _W("OpenTelemetry.Instrumentation.Wcf"), {1, 16, 0, 1139} }, \
         { _W("OpenTelemetry.OpAmp.Client"), {0, 4, 0, 1087} }, \
         { _W("OpenTelemetry.Resources.Azure"), {1, 15, 1, 1031} }, \
         { _W("OpenTelemetry.Resources.Container"), {1, 15, 1, 1038} }, \